Make graphql a peer dependency
The graphql package does not work properly when it is included multiple times in an application.

Since Pathfinder is bundled, all control of the graphql package is out of the hands of the end user. If they try to import it themselves they will get an error thrown that there are multiple instances of graphql installed, and there's no way around it with overrides or other conventional methods for resolving troubles like this because of the bundling of the dependency.

Moving graphql to a peer dependency and marking it as an external dependency to vite ensures that the end user can fully control how graphql gets installed.